Move the dialog form controls into the dialog namespace
