+++ ts/WoltLabSuite/Core/Dialog/form-control.ts
